requirements.txt
#pip install -r requirements.txt
requests
tqdm 
pyyaml
gurobipy==9.1.2
pandas
numpy
torch
torchvision
opencv-python
Pillow
seaborn
sympy
matplotlib
pyparsing